Initial commit: project organization and a small example
NOTE: Did not invest too much time searching Xilinx's information
      scarce documentation.

This seems to be the minimal set of files that allows opening
the project in Vivado_HLS. There is some state in the project
configuration files and moving some generated files (e.g. solution*)
causes the GUI to fail handling the project.

Gitignore is set up, through trial and error, to grab only the necessary
and avoid autogenerated files. This should be verified further.

Project organization:
  - base_* : common and generic code,
  - rig_*  : modules for code setup for test and synthesis
  	     (e.g. function template instantiations due to the restriction
	     on top_function not being a template),
  - bench_*: testbench modules needed due to observed restrictions on
  	     acceptable code in _Source_ section files.

The organization was chosen due to observed restrictions through
trial and error. Better setups might be possible.